move concrete array span to compute internal
